puppycompanyblog/requirements.txt
Click==7.0
Flask==1.0.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.0
Werkzeug==0.14.1
WTForms==2.2.1
flask_sqlalchemy==2.4.0
flask_migrate
flask_login
pycountry
paypalrestsdk
flask_mail
pillow
mysqlclient
